fix: fix frequency offset calculation and buffer size checks
Updated the frequency offset calculation to negate the result of the Atan2 function to correctly compute offsets. Also added a check to prevent inappropriate buffer sizes (size of 2 or less) which could lead to errors during computations. This is to ensure that there are always sufficient samples for frequency offset calculations.